refactor: move redirect routes
