+++ apps/arclight/src/app/[...route]/_s/index.ts
@@ -1,0 +1,111 @@
+import { OpenAPIHono, createRoute, z } from '@hono/zod-openapi'
+import { ResultOf, graphql } from 'gql.tada'
+import { HTTPException } from 'hono/http-exception'
+
+import { getApolloClient } from '../../../lib/apolloClient'
+import { setCorsHeaders } from '../../_redirectUtils'
+
+const GET_VIDEO_VARIANT = graphql(`
+  query GetVideoWithVariant($id: ID!, $languageId: ID!) {
+    video(id: $id) {
+      variant(languageId: $languageId) {
+        hls
+        dash
+        share
+        downloads {
+          quality
+          url
+        }
+      }
+    }
+  }
+`)
+
+const getVideoVariant = async (
+  mediaComponentId: string,
+  languageId: string
+) => {
+  try {
+    const { data } = await getApolloClient().query<
+      ResultOf<typeof GET_VIDEO_VARIANT>
+    >({
+      query: GET_VIDEO_VARIANT,
+      variables: {
+        id: mediaComponentId,
+        languageId
+      }
+    })
+    if (!data.video?.variant) {
+      throw new HTTPException(404, { message: 'Video variant not found' })
+    }
+    return data.video.variant
+  } catch (error) {
+    throw new HTTPException(500, { message: 'Failed to fetch video data' })
+  }
+}
+
+const sRoute = createRoute({
+  method: 'get',
+  path: '/:mediaComponentId/:languageId',
+  tags: ['Redirects by-convention'],
+  summary: 'Redirects to the shareable watch page',
+  description:
+    'Redirects to the shareable watch page for the given media component and language IDs.',
+  request: {
+    params: z.object({
+      mediaComponentId: z.string().describe('The ID of the media component'),
+      languageId: z.string().describe('The ID of the language')
+    })
+  },
+  responses: {
+    302: {
+      description: 'Redirects to the shareable watch page'
+    },
+    404: {
+      content: {
+        'application/json': {
+          schema: z.object({
+            error: z.string()
+          })
+        }
+      },
+      description: 'Video variant not found'
+    },
+    500: {
+      content: {
+        'application/json': {
+          schema: z.object({
+            error: z.string()
+          })
+        }
+      },
+      description: 'Internal server error'
+    }
+  }
+} as const)
+
+export const s = new OpenAPIHono()
+
+s.openapi(sRoute, async (c) => {
+  setCorsHeaders(c)
+  const { mediaComponentId, languageId } = c.req.param()
+  try {
+    const variant = await getVideoVariant(mediaComponentId, languageId)
+    if (variant.share) {
+      return c.redirect(variant.share, 302)
+    }
+    // Fallback to jesusfilm.org watch page
+    return c.redirect(
+      `http://jesusfilm.org/bin/jf/watch.html/${mediaComponentId}/${languageId}`,
+      302
+    )
+  } catch (error) {
+    if (error instanceof HTTPException) {
+      if (error.status === 404) {
+        return c.json({ error: error.message }, 404)
+      }
+    }
+    const errorMessage = error instanceof Error ? error.message : String(error)
+    return c.json({ error: `Internal server error: ${errorMessage}` }, 500)
+  }
+})
