+++ apps/arclight/src/lib/brightcove.ts
@@ -94,9 +94,7 @@ export async function getBrightcoveVideo(
     }
     if (ip) {
       apiHeaders['X-Forwarded-For'] = ip
-      console.log(`[Brightcove] Forwarding IP: ${ip}`)
     }
-    console.log('[Brightcove] fetching video', ovpReferenceId, 'with key', idx)
     const response = await fetch(
       `https://edge.api.brightcove.com/playback/v1/accounts/${ACCOUNT_ID}/videos/ref:${encodeURIComponent(
         ovpReferenceId
@@ -108,11 +106,6 @@ export async function getBrightcoveVideo(
 
     if (response.ok) {
       const json = await response.json()
-      console.log(
-        '[Brightcove] API response JSON:',
-        JSON.stringify(json, null, 2)
-      )
-      console.log('[Brightcove] success with key', idx, 'caching result')
       if (!forceRefresh) {
         setCache(ovpReferenceId, json)
       }
@@ -120,7 +113,7 @@ export async function getBrightcoveVideo(
     }
 
     if (response.status === 401 || response.status === 403) {
-      console.warn('[Brightcove] unauthorized for key', idx)
+      // console.warn('[Brightcove] unauthorized for key', idx)
     }
     if (response.status !== 401 && response.status !== 403) {
       // Errors other than unauthorized – break and throw
@@ -140,56 +133,43 @@ export function selectBrightcoveSource(
   video: any,
   code: BrightcoveSourceCode
 ): string | null {
-  console.log('[Brightcove] selectBrightcoveSource code:', code)
-  console.log(
-    '[Brightcove] video.sources:',
-    JSON.stringify(video?.sources, null, 2)
-  )
   if (!video?.sources || !Array.isArray(video.sources)) return null
   const sources = video.sources as Array<any>
-  console.log('[Brightcove] sources found:', sources)
 
   switch (code) {
     case 'hls': {
       const src = sources.find((s) => s.type === 'application/x-mpegURL')
-      console.log('[Brightcove] src selected:', src)
       return src?.src ?? null
     }
     case 'dash': {
       const src = sources.find(
         (s) => s.container === 'MPD' && s.type === 'application/dash+xml'
       )
-      console.log('[Brightcove] src selected:', src)
       return src?.src ?? null
     }
     case 'dh': {
       const mp4s = sources.filter((s) => s.container === 'MP4')
-      console.log('[Brightcove] MP4s for dh:', mp4s)
       if (!mp4s.length) return null
       const sorted = mp4s.sort(
         (a, b) => (b.avg_bitrate ?? 0) - (a.avg_bitrate ?? 0)
       )
-      console.log('[Brightcove] sorted MP4s for dh:', sorted)
       return sorted[0]?.src ?? null
     }
     case 'dl': {
       const mp4s = sources.filter((s) => s.container === 'MP4')
-      console.log('[Brightcove] MP4s for dl:', mp4s)
       if (!mp4s.length) return null
       const sorted = mp4s.sort(
         (a, b) => (a.avg_bitrate ?? 0) - (b.avg_bitrate ?? 0)
       )
-      console.log('[Brightcove] sorted MP4s for dl:', sorted)
       return sorted[0]?.src ?? null
     }
     case 's': {
       const social = sources.find((s) => /social/i.test(s.src))
-      console.log('[Brightcove] social src:', social)
       if (social?.src) return social.src
       // fallback to brightcove player page
       if (!ACCOUNT_ID || !video?.id) return null
       const fallbackUrl = `https://players.brightcove.net/${ACCOUNT_ID}/default_default/index.html?videoId=${video.id}`
-      console.log('[Brightcove] fallback player page URL:', fallbackUrl)
+      // [Brightcove] fallback player page URL: https://players.brightcove.net/[ACCOUNT_ID_REDACTED]/default_default/index.html?videoId=${video.id}
       return fallbackUrl
     }
     default:
